requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.4.0
sqlalchemy>=2.0.30
alembic>=1.13.1
pytest>=7.4.0
psycopg[binary]>=3.1.0
redis>=5.0.0
motor>=3.3.0
python-keycloak>=4.2.0
httpx>=0.27.0
PyJWT[crypto]>=2.9.0
opentelemetry-api>=1.25.0
opentelemetry-sdk>=1.25.0
opentelemetry-exporter-otlp>=1.25.0
opentelemetry-instrumentation-fastapi>=0.46b0
opentelemetry-instrumentation-sqlalchemy>=0.46b0
opentelemetry-instrumentation-redis>=0.46b0
opentelemetry-instrumentation-pymongo>=0.46b0
email-validator>=2.3.0
